+++ lib/bolt/cli.rb
@@ -132,6 +132,9 @@ module Bolt
       # Logger must be configured before checking path case, otherwise warnings will not display
       @config.check_path_case('modulepath', @config.modulepath)
 
+      # Log the file paths for loaded config files
+      @config.config_loaded
+
       parser.warnings.each { |warning| @logger.warn(warning[:msg]) }
       # After validation, initialize inventory and targets. Errors here are better to catch early.
       # After this step