+++ lib/bolt/config.rb
@@ -45,13 +45,6 @@ module Bolt
       "concurrency"              => "The number of threads to use when executing on remote targets.",
       "format"                   => "The format to use when printing results. Options are `human` and `json`.",
       "hiera-config"             => "The path to your Hiera config.",
-      "interpreters"             => "A map of an extension name to the absolute path of an executable, "\
-                                    "enabling you to override the shebang defined in a task executable. The "\
-                                    "extension can optionally be specified with the `.` character (`.py` and "\
-                                    "`py` both map to a task executable `task.py`) and the extension is case "\
-                                    "sensitive. The transports that support interpreter configuration are "\
-                                    "`docker`, `local`, `ssh`, and `winrm`. When a target's name is `localhost`, "\
-                                    "Ruby tasks run with the Bolt Ruby interpreter by default.",
       "inventoryfile"            => "The path to a structured data inventory file used to refer to groups of "\
                                     "targets on the command line and from plans.",
       "log"                      => "The configuration of the logfile output. Configuration can be set for "\
@@ -60,6 +53,8 @@ module Bolt
                                     "of directories or a string containing a list of directories separated by the "\
                                     "OS-specific PATH separator.",
       "plugin_hooks"             => "Which plugins a specific hook should use.",
+      "plugins"                  => "A map of plugins and their configuration data.",
+      "puppetdb"                 => "A map containing options for configuring the Bolt PuppetDB client.",
       "puppetfile"               => "A map containing options for the `bolt puppetfile install` command.",
       "save-rerun"               => "Whether to update `.rerun.json` in the Bolt project directory. If "\
                                     "your target names include passwords, set this value to `false` to avoid "\
@@ -119,17 +114,37 @@ module Bolt
 
     def self.from_boltdir(boltdir, overrides = {})
       data = Bolt::Util.read_config_file(nil, [boltdir.config_file], 'config') || {}
-      new(boltdir, data, overrides)
+      new(boltdir, data, overrides, load_defaults)
     end
 
     def self.from_file(configfile, overrides = {})
       boltdir = Bolt::Boltdir.new(Pathname.new(configfile).expand_path.dirname)
       data = Bolt::Util.read_config_file(configfile, [], 'config') || {}
-
-      new(boltdir, data, overrides)
+      new(boltdir, data, overrides, load_defaults)
     end
 
-    def initialize(boltdir, config_data, overrides = {})
+    def self.load_defaults
+      # Lazy-load expensive gem code
+      require 'win32/dir' if Bolt::Util.windows?
+
+      system_path = if Bolt::Util.windows?
+                      Pathname.new(File.join(Dir::COMMON_APPDATA, 'PuppetLabs', 'bolt', 'etc', 'bolt.yaml'))
+                    else
+                      Pathname.new(File.join('/etc', 'puppetlabs', 'bolt', 'bolt.yaml'))
+                    end
+      user_path = Pathname.new(File.expand_path(File.join('~', '.puppetlabs', 'etc', 'bolt', 'bolt.yaml')))
+
+      system_data = Bolt::Util.read_config_file(nil, [system_path], 'config') || {}
+      user_data = Bolt::Util.read_config_file(nil, [user_path], 'config') || {}
+
+      data = []
+      data << { filepath: system_path, data: system_data } unless system_data.empty?
+      data << { filepath: user_path, data: user_data } unless user_data.empty?
+
+      data
+    end
+
+    def initialize(boltdir, config_data, overrides = {}, defaults = [])
       @logger = Logging.logger[self]
 
       @boltdir = boltdir
@@ -143,7 +158,11 @@ module Bolt
       @puppetfile_config = {}
       @plugins = {}
       @plugin_hooks = {}
+<<<<<<< HEAD
       @apply_settings = {}
+=======
+      @config_files = []
+>>>>>>> (GH-608) Add support for multiple configuration files
 
       # add an entry for the default console logger
       @log = { 'console' => {} }
@@ -152,6 +171,11 @@ module Bolt
 
       TRANSPORTS.each do |key, transport|
         @transports[key] = transport.default_options
+      end
+
+      defaults.each do |default|
+        @config_files << default[:filepath]
+        update_from_file(default[:data])
       end
 
       update_from_file(config_data)
@@ -214,6 +238,10 @@ module Bolt
         update_logs(data['log'])
       end
 
+      if data['plugins'].is_a?(Hash)
+        update_plugins(data['plugins'])
+      end
+
       # Expand paths relative to the Boltdir. Any settings that came from the
       # CLI will already be absolute, so the expand will be skipped.
       if data.key?('modulepath')
@@ -246,13 +274,31 @@ module Bolt
 
       @save_rerun = data['save-rerun'] if data.key?('save-rerun')
 
+<<<<<<< HEAD
       %w[concurrency format puppetdb color plugins plugin_hooks].each do |key|
+=======
+      @plugin_hooks.merge!(data['plugin_hooks']) if data.key?('plugin_hooks')
+
+      %w[concurrency format puppetdb color].each do |key|
+>>>>>>> (GH-608) Add support for multiple configuration files
         send("#{key}=", data[key]) if data.key?(key)
       end
 
       update_transports(data)
     end
     private :update_from_file
+
+    # Shallow merge config for individual plugins
+    def update_plugins(data)
+      data.each_pair do |plugin, config|
+        if @plugins[plugin]
+          @plugins[plugin].merge!(config)
+        else
+          @plugins[plugin] = config
+        end
+      end
+    end
+    private :update_plugins
 
     def apply_overrides(options)
       %i[concurrency transport format trace modulepath inventoryfile color].each do |key|
@@ -401,6 +447,13 @@ module Bolt
       path.chars.map do |l|
         l =~ /[A-Za-z]/ ? "[#{l.upcase}#{l.downcase}]" : l
       end.join
+    end
+
+    def config_loaded
+      msg = <<~MSG
+        Loaded configuration from: '#{[*@config_files, boltdir.config_file].join("', '")}'
+      MSG
+      @logger.debug(msg)
     end
   end
 end