+++ CHANGELOG.md
@@ -36,6 +36,11 @@
 
   The `prompt` plugin now requires a `message` option.
 
+* **Support for system-wide and user-level configuration files** ([#608](https://github.com/puppetlabs/bolt/issues/608))
+
+  Bolt now supports system-wide and user configuration files, in addition to the existing project configuration file.
+  File precedence and merge strategy can be found in the [Bolt configuration docs](https://puppet.com/docs/bolt/latest/configuring_bolt.html).
+
 ## Bolt 1.47.0
 
 ### Deprecations and removals